+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   Rust and JSON-RPC API have `flags` integer argument
   replaced with two boolean flags `info_only` and `add_daymarker`.
 - jsonrpc: add API to check if the message is sent by a bot #3877
+- Remove `MimeMessage::from_bytes()` public interface. #4033
 
 
 ## 1.107.1